fix: use node:fs instead of `Bun.fs` in `ElysiaAdapter` (#947)
* fix(elysia): use node:fs instead of `Bun.fs`

this way we can use the same code for bun and node, as elysia is compatible to both.

(experimental) fs.glob/fs.promises.glob has been [added](https://nodejs.org/docs/latest-v22.x/api/fs.html#fspromisesglobpattern-options) in node >=22

* fix(elysia): return proper mime-types for static files

the files all had a mime-type of "", which caused loading issues in some setups.

added it with an alias (mimeV4) as not so conflict with the outdated @types/mime package in a sister package.

---------

Co-authored-by: Felix Mosheev <9304194+felixmosh@users.noreply.github.com>